+++ tests/ui/empty_parens.rs
@@ -1,0 +1,7 @@
+use diesel_derive_newtype::DieselNewType;
+
+#[derive(DieselNewType)]
+#[diesel_newtype()]
+struct EmptyParens(i64);
+
+fn main() {}