+++ tests/ui/enum.rs
@@ -1,0 +1,9 @@
+use diesel_derive_newtype::DieselNewType;
+
+#[derive(DieselNewType)]
+enum Color {
+    Red,
+    Green,
+}
+
+fn main() {}