+++ tests/ui/named_field.rs
@@ -1,0 +1,8 @@
+use diesel_derive_newtype::DieselNewType;
+
+#[derive(DieselNewType)]
+struct Named {
+    value: i64,
+}
+
+fn main() {}