+++ tests/ui/two_fields.rs
@@ -1,0 +1,6 @@
+use diesel_derive_newtype::DieselNewType;
+
+#[derive(DieselNewType)]
+struct TwoFields(i64, i64);
+
+fn main() {}