+++ tests/ui/unit_struct.rs
@@ -1,0 +1,6 @@
+use diesel_derive_newtype::DieselNewType;
+
+#[derive(DieselNewType)]
+struct Unit;
+
+fn main() {}