+++ tests/ui/zero_fields.rs
@@ -1,0 +1,6 @@
+use diesel_derive_newtype::DieselNewType;
+
+#[derive(DieselNewType)]
+struct ZeroFields();
+
+fn main() {}